// lib/models/models.type.ts
export interface JobApplication {
  _id: string;
  company: string;
  position: string;
  location?: string;
  status: string;
  notes?: string;
  salary?: number;
  jobUrl?: string;
  order: number;
  columnId: string;
  tags?: string[];
  description?: string;
}

export interface column {
  _id: string;
  name: string;
  order: number;
  jobApplications: JobApplication[];
}

export interface Board {
  _id: string;
  name: string;
  columns: column[];
}